moving tipper to its own module
