+++ mtpy/core/z.py
@@ -11,22 +11,19 @@ Updated 11/2020 for logging and formating (J. Peacock).
     - ToDo: add functionality for covariance matrix
 """
 
-# =================================================================
-import cmath
+# =============================================================================
+# Imports
+# =============================================================================
 import copy
-import math
 
 import numpy as np
 
 import mtpy.utils.calculator as MTcc
 from mtpy.utils.exceptions import (
     MTpyError_Z,
-    MTpyError_Tipper,
-    MTpyError_input_arguments,
 )
-from mtpy.utils.mtpy_logger import get_mtpy_logger
-from .res_phase import ResPhase
 
+from .res_phase import ResPhase
 
 # ==============================================================================
 # Impedance Tensor Class
@@ -94,9 +91,6 @@ class Z(ResPhase):
 
         """
         self.rotation_angle = 0.0
-        self._z = None
-        self._z_err = None
-        self._z_model_err = None
 
         super().__init__()
 
@@ -293,6 +287,7 @@ class Z(ResPhase):
 
         Nulling the rotation_angle
         """
+
         old_shape = None
         if self._z is not None:
             old_shape = self._z.shape
@@ -318,6 +313,7 @@ class Z(ResPhase):
                             deviation
         :type z_err_array: np.ndarray(nfrequency, 2, 2)
         """
+
         old_shape = None
         if self._z is not None:
             old_shape = self._z.shape
@@ -337,9 +333,10 @@ class Z(ResPhase):
         Set the attribute z_model_err
 
         :param z_model_err_array: error of impedance tensor array as standard
-                            deviation
+         deviation
         :type z_model_err_array: np.ndarray(nfrequency, 2, 2)
         """
+
         old_shape = None
         if self._z is not None:
             old_shape = self._z.shape
@@ -929,741 +926,3 @@ class Z(ResPhase):
             )
 
         return invariants_dict
-
-
-# ======================================================================
-#                               TIPPER
-# ======================================================================
-class Tipper(object):
-    """
-    Tipper class --> generates a Tipper-object.
-
-    Errors are given as standard deviations (sqrt(VAR))
-
-    :param tipper_array: tipper array in the shape of [Tx, Ty]
-                         *default* is None
-    :type tipper_array: np.ndarray((nf, 1, 2), dtype='complex')
-
-
-    :param tipper_err_array: array of estimated tipper errors
-                             in the shape of [Tx, Ty].
-                             Must be the same shape as tipper_array.
-                             *default* is None
-    :type tipper_err_array: np.ndarray((nf, 1, 2))
-
-
-    :param frequency: array of frequencyuencies corresponding to the tipper elements.
-                 Must be same length as tipper_array.
-                 *default* is None
-    :type frequency: np.ndarray(nf)
-
-
-    =============== ===========================================================
-    Attributes      Description
-    =============== ===========================================================
-    frequency            array of frequencyuencies corresponding to elements of z
-    rotation_angle  angle of which data is rotated by
-
-    tipper          tipper array
-    tipper_err       tipper error array
-    =============== ===========================================================
-
-    =============== ===========================================================
-    Methods         Description
-    =============== ===========================================================
-    mag_direction   computes magnitude and direction of real and imaginary
-                    induction arrows.
-    amp_phase       computes amplitude and phase of Tx and Ty.
-    rotate          rotates the data by the given angle
-    =============== ===========================================================
-    """
-
-    def __init__(
-        self, tipper_array=None, tipper_err_array=None, frequency=None
-    ):
-        """
-        initialize
-        """
-        self.logger = get_mtpy_logger(self.__class__.__name__)
-        self._tipper = tipper_array
-        self._tipper_err = tipper_err_array
-        self._frequency = frequency
-
-        self.rotation_angle = 0.0
-        if self.tipper is not None:
-            self.rotation_angle = np.zeros((len(self.tipper)))
-        self._amplitude = None
-        self._amplitude_err = None
-        self._phase = None
-        self._phase_err = None
-
-        self._mag_real = None
-        self._mag_imag = None
-        self._angle_real = None
-        self._angle_imag = None
-        self._mag_err = None
-        self._angle_err = None
-
-        if self._tipper is not None and self._frequency is not None:
-            self.compute_amp_phase()
-            self.compute_mag_direction()
-
-    def __str__(self):
-        lines = ["Induction Vector (Tippers)", "-" * 40]
-        if self.frequency is not None:
-            lines.append(
-                f"\tNumber of frequencyuencies:  {self.frequency.size}"
-            )
-            lines.append(
-                f"\tfrequencyuency range:        {self.frequency.min():.5E} -- {self.frequency.max():.5E} Hz"
-            )
-            lines.append(
-                f"\tPeriod range:           {1/self.frequency.max():.5E} -- {1/self.frequency.min():.5E} s"
-            )
-            lines.append("")
-            lines.append("\tElements:")
-            for zz, ff in zip(self.tipper, self.frequency):
-                lines.append(f"\tfrequencyuency: {ff:5E} Hz -- Period {1/ff} s")
-                lines.append(
-                    "\t\t"
-                    + np.array2string(
-                        zz,
-                        formatter={
-                            "complex_kind": lambda x: f"{x.real:.4e}{x.imag:+.4E}"
-                        },
-                    ).replace("\n", "\n\t\t")
-                )
-        else:
-            if self.tipper is not None:
-                lines.append("Elements:")
-                for ff, zz in enumerate(self.tipper):
-                    lines.append(f"\tIndex {ff}")
-                    lines.append(
-                        "\t\t"
-                        + np.array2string(
-                            zz,
-                            formatter={
-                                "complex_kind": lambda x: f"{x.real:.4e}{x.imag:+.4E}"
-                            },
-                        ).replace("\n", "\n\t\t")
-                    )
-        return "\n".join(lines)
-
-    def __repr__(self):
-        return self.__str__()
-
-    def __eq__(self, other):
-        if not isinstance(other, Tipper):
-            msg = f"Cannot compare {type(other)} with Tipper"
-            self.logger.error(msg)
-            raise MTpyError_Tipper(msg)
-        if (self.tipper != other.tipper).all():
-            return False
-        if (self.frequency != other.frequency).all():
-            return False
-        if (self.tipper_err != other.tipper_err).all():
-            return False
-        return True
-
-    def copy(self):
-        return copy.deepcopy(self)
-
-    # ==========================================================================
-    # Define get/set and properties
-    # ==========================================================================
-    # ----frequency----------------------------------------------------------
-    @property
-    def frequency(self):
-        return self._frequency
-
-    @frequency.setter
-    def frequency(self, frequency_arr):
-        """
-        Set the array of frequency.
-
-        :param frequency_arr: array of frequencyunecies (Hz)
-        :type frequency_arr: np.ndarray(num_frequencyuencies)
-        """
-        if frequency_arr is None:
-            return
-        self._frequency = np.array(frequency_arr, dtype="float")
-
-        if self.tipper is not None:
-            if self.tipper.shape[0] != len(self._frequency):
-                msg = (
-                    "New frequency array is not correct shape for existing z"
-                    + "new: {self._frequency.size} != old: {self.tipper.shape[0]}"
-                )
-                self.logger.error(msg)
-                raise MTpyError_Tipper
-        # for consistency recalculate amplitude and phase
-        self.compute_amp_phase()
-
-    @property
-    def period(self):
-        """
-        periods in seconds
-        """
-
-        return 1.0 / self.frequency
-
-    @period.setter
-    def period(self, value):
-        """
-        setting periods will set the frequencyuencies
-        """
-
-        self.frequency = 1.0 / value
-
-    # ---tipper--------------------------------------------------------------
-    @property
-    def tipper(self):
-        return self._tipper
-
-    @tipper.setter
-    def tipper(self, tipper_array):
-        """
-        Set the attribute *tipper*
-
-        :param tipper_array: tipper array in the shape of [Tx, Ty]
-                             *default* is None
-        :type tipper_array: np.ndarray((nf, 1, 2), dtype='complex')
-        """
-        if tipper_array is None:
-            return
-        if not isinstance(tipper_array, np.ndarray):
-            tipper_array = np.array(tipper_array, dtype="complex")
-        if not tipper_array.dtype in ["complex"]:
-            tipper_array = tipper_array.astype("complex")
-        # check to see if the new tipper array is the same shape as the old
-        if (
-            self._tipper is not None
-            and self._tipper.shape != tipper_array.shape
-        ):
-            msg = (
-                "Shape of new array does not match old.  "
-                + f"new shape {tipper_array.shape} != "
-                + f"old shape {self._tipper.shape}. "
-                + "Make a new Tipper instance to be save."
-            )
-            self.logger.error(msg)
-            raise MTpyError_Tipper(msg)
-        if len(tipper_array.shape) == 3:
-            if tipper_array.shape[1:3] == (1, 2):
-                self._tipper = tipper_array
-            else:
-                msg = f"Input array must be shape (n, 1, 2) not {tipper_array.shape}"
-                self.logger.error(msg)
-                raise MTpyError_Tipper(msg)
-        elif len(tipper_array.shape) == 2:
-            if tipper_array.shape == (1, 2):
-                self._tipper = tipper_array.reshape((1, 1, 2))
-                self.logger.debug(
-                    "setting input tipper with shape (1, 2) to (1, 1, 2)"
-                )
-            else:
-                msg = f"Input array must be shape (n, 1, 2) not {tipper_array.shape}"
-                self.logger.error(msg)
-                raise MTpyError_Tipper(msg)
-        else:
-            msg = f"{tipper_array.shape} are not the correct dimensions, must be (n, 1, 2)"
-            self.logger.error(msg)
-            raise MTpyError_Tipper(msg)
-        # neeed to set the rotation angle such that it is an array
-        if self.rotation_angle is float:
-            self.rotation_angle = np.repeat(
-                self.rotation_angle, len(self._tipper)
-            )
-        # for consistency recalculate mag and angle
-        self.compute_mag_direction()
-
-        # for consistency recalculate amplitude and phase
-        self.compute_amp_phase()
-
-    # ----tipper error---------------------------------------------------------
-    @property
-    def tipper_err(self):
-        return self._tipper_err
-
-    @tipper_err.setter
-    def tipper_err(self, tipper_err_array):
-        """
-        Set the attribute *tipper_err*.
-
-        :param tipper_err_array: array of estimated tipper errors
-                                 in the shape of [Tx, Ty].
-                                 Must be the same shape as tipper_array.
-                                 *default* is None
-        :type tipper_err_array: np.ndarray((nf, 1, 2))
-        """
-        if tipper_err_array is None:
-            return
-        if not isinstance(tipper_err_array, np.ndarray):
-            tipper_err_array = np.array(tipper_err_array, dtype="float")
-        if not tipper_err_array.dtype in ["float"]:
-            tipper_err_array = tipper_err_array.astype("float")
-        if len(tipper_err_array.shape) == 3:
-            if not tipper_err_array.shape[1:3] == (1, 2):
-                msg = f"Input array must be shape (n, 1, 2) not {tipper_err_array.shape}"
-                self.logger.error(msg)
-                raise MTpyError_Tipper(msg)
-        elif len(tipper_err_array.shape) == 2:
-            if tipper_err_array.shape == (1, 2):
-                tipper_err_array = tipper_err_array.reshape((1, 1, 2))
-                self.logger.debug(
-                    "setting input tipper with shape (1, 2) to (1, 1, 2)"
-                )
-            else:
-                msg = f"Input array must be shape (n, 1, 2) not {tipper_err_array.shape}"
-                self.logger.error(msg)
-                raise MTpyError_Tipper(msg)
-        else:
-            msg = f"{tipper_err_array.shape} are not the correct dimensions, must be (n, 1, 2)"
-            self.logger.error(msg)
-            raise MTpyError_Tipper(msg)
-        # check to see if the new tipper array is the same shape as the old
-        if (
-            self._tipper is not None
-            and self._tipper.shape != tipper_err_array.shape
-        ):
-            raise MTpyError_Tipper(
-                "Shape of new error array does not match old"
-                + f"new shape {tipper_err_array.shape} != old shape {self._tipper.shape}"
-            )
-        self._tipper_err = tipper_err_array
-
-        # for consistency recalculate mag and angle
-        self.compute_mag_direction()
-
-        # for consistency recalculate amplitude and phase
-        self.compute_amp_phase()
-
-    # ----amplitude and phase
-    def compute_amp_phase(self):
-        """
-        Sets attributes:
-                        * *amplitude*
-                        * *phase*
-                        * *amplitude_err*
-                        * *phase_err*
-
-        values for resistivity are in in Ohm m and phase in degrees.
-        """
-
-        if self.tipper is None:
-            # logging.error( 'tipper array is None - cannot calculate rho/phi')
-            # print 'tipper array is None - cannot calculate rho/phi'
-            return None
-        self._amplitude_err = None
-        self._phase_err = None
-        if self.tipper_err is not None:
-            self._amplitude_err = np.zeros(self.tipper_err.shape)
-            self._phase_err = np.zeros(self.tipper_err.shape)
-        self._amplitude = np.abs(self.tipper)
-        self._phase = np.rad2deg(np.angle(self.tipper))
-
-        if self.tipper_err is not None:
-            for idx_f in range(len(self.tipper)):
-                for jj in range(2):
-                    if self.tipper_err is not None:
-                        if type(self.tipper) == np.ma.core.MaskedArray:
-                            if self.tipper.mask[idx_f, 0, jj]:
-                                continue
-                        r_err, phi_err = MTcc.propagate_error_rect2polar(
-                            np.real(self.tipper[idx_f, 0, jj]),
-                            self.tipper_err[idx_f, 0, jj],
-                            np.imag(self.tipper[idx_f, 0, jj]),
-                            self.tipper_err[idx_f, 0, jj],
-                        )
-
-                        self.amplitude_err[idx_f, 0, jj] = r_err
-                        self._phase_err[idx_f, 0, jj] = phi_err
-
-    def set_amp_phase(self, r_array, phi_array):
-        """
-        Set values for amplitude(r) and argument (phi - in degrees).
-
-        Updates the attributes:
-                        * tipper
-                        * tipper_err
-
-        """
-
-        if self.tipper is not None:
-
-            tipper_new = copy.copy(self.tipper)
-
-            if self.tipper.shape != r_array.shape:
-                self.logger.error(
-                    'Error - shape of "r" array does not match shape of '
-                    + "tipper array: %s ; %s"
-                    % (str(r_array.shape), str(self.tipper.shape))
-                )
-                return
-            if self.tipper.shape != phi_array.shape:
-                self.logger.error(
-                    'Error - shape of "phi" array does not match shape of '
-                    + "tipper array: %s ; %s"
-                    % (str(phi_array.shape), str(self.tipper.shape))
-                )
-                return
-        else:
-
-            tipper_new = np.zeros(r_array.shape, "complex")
-
-            if r_array.shape != phi_array.shape:
-                self.logger.error(
-                    'Error - shape of "phi" array does not match shape '
-                    + 'of "r" array: %s ; %s'
-                    % (str(phi_array.shape), str(r_array.shape))
-                )
-                return
-        # assert real array:
-        if np.linalg.norm(np.imag(r_array)) != 0:
-            self.logger.error('Error - array "r" is not real valued !')
-            return
-        if np.linalg.norm(np.imag(phi_array)) != 0:
-            self.logger.error('Error - array "phi" is not real valued !')
-            return
-        for idx_f in range(len(r_array)):
-            for jj in range(2):
-                tipper_new[idx_f, 0, jj] = cmath.rect(
-                    r_array[idx_f, 0, jj],
-                    math.radians(phi_array[idx_f, 0, jj]),
-                )
-        self.tipper = tipper_new
-
-        # for consistency recalculate amplitude and phase
-        self.compute_amp_phase()
-        self.compute_mag_direction()
-
-    # ---------------------------------
-    # properties
-    @property
-    def amplitude(self):
-        return self._amplitude
-
-    @property
-    def phase(self):
-        return self._phase
-
-    @property
-    def amplitude_err(self):
-        return self._amplitude_err
-
-    @property
-    def phase_err(self):
-        return self._phase_err
-
-    # ----magnitude and direction----------------------------------------------
-    def compute_mag_direction(self):
-        """
-        computes the magnitude and direction of the real and imaginary
-        induction vectors.
-        """
-
-        if self.tipper is None:
-            return None
-        self._mag_real = np.sqrt(
-            self.tipper[:, 0, 0].real ** 2 + self.tipper[:, 0, 1].real ** 2
-        )
-        self._mag_imag = np.sqrt(
-            self.tipper[:, 0, 0].imag ** 2 + self.tipper[:, 0, 1].imag ** 2
-        )
-
-        self._mag_err = None
-        self._angle_err = None
-        # get the angle, need to make both parts negative to get it into the
-        # parkinson convention where the arrows point towards the conductor
-
-        self._angle_real = np.rad2deg(
-            np.arctan2(-self.tipper[:, 0, 1].real, -self.tipper[:, 0, 0].real)
-        )
-
-        self._angle_imag = np.rad2deg(
-            np.arctan2(-self.tipper[:, 0, 1].imag, -self.tipper[:, 0, 0].imag)
-        )
-
-        ## estimate error: THIS MAYBE A HACK
-        if self.tipper_err is not None:
-            self._mag_err = np.sqrt(
-                self.tipper_err[:, 0, 0] ** 2 + self.tipper_err[:, 0, 1] ** 2
-            )
-            self._angle_err = (
-                np.rad2deg(
-                    np.arctan2(
-                        self.tipper_err[:, 0, 0], self.tipper_err[:, 0, 1]
-                    )
-                )
-                % 45
-            )
-
-    def set_mag_direction(self, mag_real, ang_real, mag_imag, ang_imag):
-        """
-        computes the tipper from the magnitude and direction of the real
-        and imaginary components.
-
-        Updates tipper
-
-        No error propagation yet
-        """
-
-        self.tipper[:, 0, 0].real = np.sqrt(
-            (mag_real**2 * np.arctan(ang_real) ** 2)
-            / (1 - np.arctan(ang_real) ** 2)
-        )
-
-        self.tipper[:, 0, 1].real = np.sqrt(
-            mag_real**2 / (1 - np.arctan(ang_real) ** 2)
-        )
-
-        self.tipper[:, 0, 0].imag = np.sqrt(
-            (mag_imag**2 * np.arctan(ang_imag) ** 2)
-            / (1 - np.arctan(ang_imag) ** 2)
-        )
-
-        self.tipper[:, 0, 1].imag = np.sqrt(
-            mag_imag**2 / (1 - np.arctan(ang_imag) ** 2)
-        )
-        # for consistency recalculate mag and angle
-        self.compute_mag_direction()
-        self.compute_amp_phase()
-
-    @property
-    def mag_real(self):
-        return self._mag_real
-
-    @property
-    def mag_imag(self):
-        return self._mag_imag
-
-    @property
-    def angle_real(self):
-        return self._angle_real
-
-    @property
-    def angle_imag(self):
-        return self._angle_imag
-
-    @property
-    def mag_err(self):
-        return self._mag_err
-
-    @property
-    def angle_err(self):
-        return self._angle_err
-
-    # ----rotate---------------------------------------------------------------
-    def rotate(self, alpha):
-        """
-        Rotate  Tipper array.
-
-        Rotation angle must be given in degrees. All angles are referenced
-        to geographic North=0, positive in clockwise direction.
-        (Mathematically negative!)
-
-        In non-rotated state, 'X' refs to North and 'Y' to East direction.
-
-        Updates the attributes:
-                        * *tipper*
-                        * *tipper_err*
-                        * *rotation_angle*
-
-        """
-
-        if self.tipper is None:
-            self.logger.error('tipper array is "None" - I cannot rotate that')
-            return
-        # check for iterable list/set of angles - if so, it must have length 1
-        # or same as len(tipper):
-        if np.iterable(alpha) == 0:
-            try:
-                degreeangle = float(alpha % 360)
-            except ValueError:
-                self.logger.error('"Angle" must be a valid number (in degrees)')
-                return
-            # make an n long list of identical angles
-            lo_angles = [degreeangle for ii in self.tipper]
-        elif len(alpha) == 1:
-            try:
-                degreeangle = float(alpha % 360)
-            except ValueError:
-                self.logger.error('"Angle" must be a valid number (in degrees)')
-                return
-            # make an n long list of identical angles
-            lo_angles = [degreeangle for ii in self.tipper]
-        else:
-            try:
-                lo_angles = [float(ii % 360) for ii in alpha]
-            except ValueError:
-                self.logger.error('"Angles" must be valid numbers (in degrees)')
-                return
-        self.rotation_angle = np.array(
-            [
-                (oldangle + lo_angles[ii]) % 360
-                for ii, oldangle in enumerate(self.rotation_angle)
-            ]
-        )
-
-        if len(lo_angles) != len(self.tipper):
-            self.logger.error(
-                'Wrong number Number of "angles" - need %ii '
-                % (len(self.tipper))
-            )
-            self.rotation_angle = 0.0
-            return
-        tipper_rot = copy.copy(self.tipper)
-        tipper_err_rot = copy.copy(self.tipper_err)
-
-        for idx_frequency in range(len(tipper_rot)):
-            angle = lo_angles[idx_frequency]
-
-            if self.tipper_err is not None:
-                (
-                    tipper_rot[idx_frequency],
-                    tipper_err_rot[idx_frequency],
-                ) = MTcc.rotate_vector_with_errors(
-                    self.tipper[idx_frequency, :, :],
-                    angle,
-                    self.tipper_err[idx_frequency, :, :],
-                )
-            else:
-                (
-                    tipper_rot[idx_frequency],
-                    tipper_err_rot,
-                ) = MTcc.rotate_vector_with_errors(
-                    self.tipper[idx_frequency, :, :], angle
-                )
-        self.tipper = tipper_rot
-        self.tipper_err = tipper_err_rot
-
-        # for consistency recalculate mag and angle
-        self.compute_mag_direction()
-
-        # for consistency recalculate amplitude and phase
-        self.compute_amp_phase()
-
-
-# ------------------------
-def correct4sensor_orientation(
-    Z_prime, Bx=0, By=90, Ex=0, Ey=90, Z_prime_error=None
-):
-    """
-    Correct a Z-array for wrong orientation of the sensors.
-
-    Assume, E' is measured by sensors orientated with the angles
-        E'x: a
-        E'y: b
-
-    Assume, B' is measured by sensors orientated with the angles
-        B'x: c
-        B'y: d
-
-    With those data, one obtained the impedance tensor Z':
-        E' = Z' * B'
-
-    Now we define change-of-basis matrices T,U so that
-        E = T * E'
-        B = U * B'
-
-    =>   T contains the expression of the E'-basis in terms of E
-    (the standard basis)
-    and  U contains the expression of the B'-basis in terms of B
-    (the standard basis)
-    The respective expressions for E'x-basis vector and E'y-basis
-    vector are the columns of T.
-    The respective expressions for B'x-basis vector and B'y-basis
-    vector are the columns of U.
-
-    We obtain the impedance tensor in default coordinates as:
-
-    E' = Z' * B' => T^(-1) * E = Z' * U^(-1) * B
-                 => E = T * Z' * U^(-1) * B
-                 => Z = T * Z' * U^(-1)
-
-    :param Z_prime: impedance tensor to be adjusted
-    :dtype Z_prime: np.ndarray(num_frequency, 2, 2, dtype='complex')
-
-
-    :param Bx: orientation of Bx relative to geographic north (0)
-                                   *default* is 0
-    :type Bx: float (angle in degrees)
-
-    :param By:
-    :type By: float (angle in degrees)
-                         orientation of By relative to geographic north (0)
-                                 *default* is 90
-
-    :param Ex: orientation of Ex relative to geographic north (0)
-                                   *default* is 0
-    :type Ex: float (angle in degrees)
-
-    :param Ey: orientation of Ey relative to geographic north (0)
-                                  *default* is 90
-    :type Ey: float (angle in degrees)
-
-    :param Z_prime_error: impedance tensor error (std)
-                                                 *default* is None
-    :type Z_prime_error: np.ndarray(Z_prime.shape)
-
-    :returns: adjusted impedance tensor
-    :rtype: np.ndarray(Z_prime.shape, dtype='complex')
-
-    :returns: impedance tensor standard deviation in
-                                        default orientation
-    :rtype: np.ndarray(Z_prime.shape, dtype='real')
-    """
-    try:
-        if len(Z_prime.shape) != 2:
-            raise
-        if Z_prime.shape != (2, 2):
-            raise
-        if Z_prime.dtype not in ["complex", "float", "int"]:
-            raise
-        Z_prime = np.matrix(Z_prime)
-    except:
-        raise MTpyError_input_arguments(
-            "ERROR - Z array not valid!" + "Must be 2x2 complex array"
-        )
-    if Z_prime_error is not None:
-        try:
-            if len(Z_prime_error.shape) != 2:
-                raise
-            if Z_prime_error.shape != (2, 2):
-                raise
-            if Z_prime_error.dtype not in ["float", "int"]:
-                raise
-        except:
-            raise MTpyError_input_arguments(
-                "ERROR - Z-error array not" + "valid! Must be 2x2 real array"
-            )
-    T = np.matrix(np.zeros((2, 2)))
-    U = np.matrix(np.zeros((2, 2)))
-
-    dummy1 = cmath.rect(1, math.radians(Ex))
-
-    T[0, 0] = np.real(dummy1)
-    T[1, 0] = np.imag(dummy1)
-    dummy2 = cmath.rect(1, math.radians(Ey))
-    T[0, 1] = np.real(dummy2)
-    T[1, 1] = np.imag(dummy2)
-
-    dummy3 = cmath.rect(1, math.radians(Bx))
-    U[0, 0] = np.real(dummy3)
-    U[1, 0] = np.imag(dummy3)
-    dummy4 = cmath.rect(1, math.radians(By))
-    U[0, 1] = np.real(dummy4)
-    U[1, 1] = np.imag(dummy4)
-
-    try:
-        z_arr = np.array(np.dot(T, np.dot(Z_prime, U.I)))
-    except:
-        raise MTpyError_input_arguments(
-            "ERROR - Given angles do not"
-            + "define basis for 2 dimensions - cannot convert Z'"
-        )
-    z_err_arr = copy.copy(Z_prime_error)
-
-    # TODO: calculate error propagation
-
-    return z_arr, z_err_arr
